Allow already formatted build IDs
Currently build ids have to be a byte slice, which makes sense if you
are obtaining them from the section where they are stored in the ELF
files. As the debuginfod servers required the formatted representation,
`fetch_debug_info` formats them before issuing the HTTP request.

This is inconvenient for projects where the formatted build id is
already stored, as we would have to convert back to raw bytes and this
crate would once again convert to the formatted representation, so this
commit generalises build ids to their own enum.

Test Plan
=========

Added the already formatted build id variant to the integration tests.