Rename BuildId::formatted() to format
It's somewhat common practice to have a convenience constructor for enum
variants named by lower casing the variant name itself. Currently this
name is already used by the BuildId::formatted() method, which really is
a getter.
Let's rename it to BuildId::format() to free up this name. That's also
closer to std::fmt::Display::fmt() in that it is a verb.