Add `Response` struct and return it in `Client::fetch_debug_info()`
Which allows for a richer set of informatiomn to be passed
to the caller beyond the actual debug information reader. The
first additional field is the debuginfod server that returned
the response. This is very helpful for systems that want to store
the source of said debug information. This is typicall used for
later debugging, re-downloading the debug information, if needed,
etc.

Test Plan
=========

Adapted the tests.